monitor test

+++ nemu/src/memory/memory.c
@@ -44,8 +44,8 @@ uint32_t laddr_read(laddr_t laddr, size_t len) {
 	return paddr_read(laddr, len);
 #else
 	if (cpu.cr0.PE == 1 && cpu.cr0.PG == 1) {
-		int start = (laddr >> 12) & 0x1;
-		int end = ((laddr + len) >> 12) & 0x1;
+		uint8_t start = (laddr >> 12) & 0x1;
+		uint8_t end = ((laddr + len) >> 12) & 0x1;
 		if (start != end) {
 			assert(0);
 		} else {
@@ -64,8 +64,8 @@ void laddr_write(laddr_t laddr, size_t len, uint32_t data) {
 #else
 	if (cpu.cr0.PE == 1 && cpu.cr0.PG == 1) {
 		// the following code is used to tell if we cross the page
-		int start = (laddr >> 12) & 0x1;
-		int end = ((laddr + len) >> 12) & 0x1;
+		uint8_t start = (laddr >> 12) & 0x1;
+		uint8_t end = ((laddr + len) >> 12) & 0x1;
 		if (start != end) {
 			assert(0);
 		} else {
